+++ inc/bs-control.php
@@ -9,23 +9,10 @@ if ( !defined( 'BS_PATH' ) ) {
 ---------------------------------------------------------- */
 
 include BS_INC_DIR . 'functions.php';
+include BS_INC_DIR . 'classes/bs-model.php';
 include BS_INC_DIR . 'classes/bs-database.php';
 include BS_INC_DIR . 'classes/bs-user.php';
 include BS_INC_DIR . 'classes/bs-page.php';
-
-/* ----------------------------------------------------------
-  Config
----------------------------------------------------------- */
-
-if ( file_exists( BS_PATH . 'bs-config.php' ) ) {
-    include BS_PATH . 'bs-config.php';
-}
-else {
-    include BS_INC_DIR . 'bs-install.php';
-    die;
-}
-
-include BS_INC_DIR . 'bs-config.php';
 
 /* ----------------------------------------------------------
   Launch project
@@ -38,10 +25,23 @@ $m = $p = 'index';
 
 if ( isset( $_GET['p'] ) ) {
     $p = '404';
-    if ( $_GET['p'] != 'index' && preg_match( '/^([a-z0-9-_]+)$/', $_GET['p'] ) ) {
+    if ( !in_array( $_GET['p'], array( 'index', 'install' ) ) && preg_match( '/^([a-z0-9-_]+)$/', $_GET['p'] ) ) {
         $m = $p = $_GET['p'];
     }
 }
+
+/* ----------------------------------------------------------
+  Config
+---------------------------------------------------------- */
+
+if ( file_exists( BS_PATH . 'bs-config.php' ) ) {
+    include BS_PATH . 'bs-config.php';
+}
+else {
+    $p = $m = 'install';
+}
+
+include BS_INC_DIR . 'bs-config.php';
 
 /* Set Model
 -------------------------- */