+++ inc/models/index.php
@@ -1,6 +1,6 @@
 <?php
 
-class BS_Model_Index {
+class BS_Model_Index extends BS_Model {
     function __construct() {
     }
 }